refactor(runner): absorb sentinel, progress, and learnings parsing into runner
Move generateNonce, detectCompletion, extractNoncedBlock from sentinel.ts,
appendProgressBlock from progress.ts, and parseLearningContent from
learnings.ts into runner.ts. These were single-consumer helpers used only
by the runner loop.

Delete sentinel.ts, sentinel.test.ts, progress.ts. Slim learnings.ts to
export only formatting functions. Update ARCHITECTURE.md to reflect the
absorbed modules.
