+++ src/learnings.test.ts
@@ -1,62 +1,5 @@
 import { describe, it, expect } from "bun:test";
 import { formatLearningsForPrompt, formatLearningsForPr } from "./learnings.ts";
-import { parseLearningContent } from "./runner.ts";
-
-// ---------------------------------------------------------------------------
-// parseLearningContent
-// ---------------------------------------------------------------------------
-
-describe("parseLearningContent", () => {
-  it('returns null for "none" (lowercase)', () => {
-    expect(parseLearningContent("none")).toBeNull();
-  });
-
-  it('returns null for "None" (mixed case)', () => {
-    expect(parseLearningContent("None")).toBeNull();
-  });
-
-  it('returns null for "NONE" (uppercase)', () => {
-    expect(parseLearningContent("NONE")).toBeNull();
-  });
-
-  it('returns null for "none" with surrounding whitespace', () => {
-    expect(parseLearningContent("  none  ")).toBeNull();
-  });
-
-  it("returns null for empty string", () => {
-    expect(parseLearningContent("")).toBeNull();
-  });
-
-  it("returns null for whitespace-only string", () => {
-    expect(parseLearningContent("   \n\t  ")).toBeNull();
-  });
-
-  it("returns trimmed prose for non-none content", () => {
-    const result = parseLearningContent(
-      "  The build fails when running on Windows due to path separators.  ",
-    );
-    expect(result).toBe(
-      "The build fails when running on Windows due to path separators.",
-    );
-  });
-
-  it("returns multiline prose as trimmed text", () => {
-    const block = [
-      "",
-      "  First line of learning.",
-      "  Second line of learning.",
-      "",
-    ].join("\n");
-    const result = parseLearningContent(block);
-    expect(result).toBe("First line of learning.\n  Second line of learning.");
-  });
-
-  it("preserves internal structure of prose", () => {
-    const block =
-      "Always use path.join() for cross-platform paths.\nNever hardcode / separators.";
-    expect(parseLearningContent(block)).toBe(block);
-  });
-});
 
 // ---------------------------------------------------------------------------
 // formatLearningsForPrompt
